6: Model Validation
Model validation is that assesses the accuracy, quality, and reliability of predicted 3D protein structure. It ensures the model is structurally sound and biologically meaningful before use it for further research or analysis.
ERRAT is a structure validation tool used to assess the overall quality of a protein model by analyzing non-bonded atomic interactions.
PROCHECK is a widely used tool for evaluating the stereochemical quality of protein structures. It focuses on whether the model’s geometry adheres to what is typically observed in high-resolution, experimentally determined structures.